+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,7 +1,7 @@
 Version x.x.x
 =============
 
- * 2015-06-30 Injection of params and uploaded files from multipart requests (bazi)
+ * 2015-07-20 Injection of params and uploaded files from multipart requests (bazi)
 
 Version 5.1.4
 =============